Input element does not fire scroll event
When there is text that overflows an input element, the scroll event is not
fired when the text scrolls into view.
The input element never receives the scroll event from the user agent shadow
DOM internal editor element. The fix is to change the scroll event's target to
the shadow host element when the event is generated.

Bug: 989937
Change-Id: Iff4aeae6adb931d681a2ade72eb8032c5a23067c