Don't make anchor clicks cancel form submission when href=#
In http://crrev.com/c/1922699 I made anchor clicks cancel all form
submissions on a page in order to make this test pass:
external/wpt/html/browsers/browsing-the-web/navigating-across-documents/006.html

The test basically says that anchor navigations should take precedence
over form submissions when the form submission is initiated by the
onclick handler of the anchor element.

However, when the anchor has href="#", form submissions should actually
go through, and this has introduced a bug on websites which rely on this
behavior.

This logic I added to cancel form submissions would not be necessary if
we made anchor clicks navigate asynchronously like the spec says to do,
so I made a bug to do so and remove this logic: http://crbug.com/1053679

Bug: 1053312
Change-Id: I27bf193cf0d544222cddad932728a1d0f61ba62a